requirements.txt
Django==3.0.4
djangorestframework==3.11.0
djongo==1.3.1
drf-yasg==1.17.1
